requirements.txt
certifi==2020.12.5
chardet==4.0.0
falcon==2.0.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
urllib3==1.26.2
